Fix entity create route. parse role refs in EntityIntermediate
